API

API

Quick Start

The easiest way to work with Meshing is with GeometryTypes. This package extends the mesh constructors from GeometryTypes.jl for convience.

The algorithms operate on a Function, AbstractArray, or SignedDistanceField and output a concrete AbstractMesh. For example:

using Meshing
using GeometryTypes
using LinearAlgebra: dot, norm
using FileIO

# Mesh an equation of sphere in the Axis-Aligned Bounding box starting
# at -1,-1,-1 and widths of 2,2,2 using Marching Cubes
m = GLNormalMesh(HyperRectangle(Vec(-1,-1,-1.), Vec(2,2,2.)), MarchingCubes()) do v
    sqrt(sum(dot(v,v))) - 1
end

# save the Sphere as a PLY file
save("sphere.ply",m)

For a full listing of concrete AbstractMesh types see GeometryTypes.jl mesh documentation.

Meshing Algorithms

Three meshing algorithms exist:

MarchingCubes()

MarchingTetrahedra()

NaiveSurfaceNets()

Each takes an optional iso and eps parameter, e.g. MarchingCubes(0.0,1e-6).

Here iso controls the offset for the boundary detection. By default this is set to 0. eps is the detection tolerance for a voxel edge intersection.

Users must construct an algorithm type and use it as an argument to a GeometryTypes mesh call or isosurface call.

Below is a comparison of the algorithms:

| | Face Type | Unique Vertices | Performance | Interpolation | |––––––––––-|–––––-|––––––––-|––––––-|–––––––––-| | Naive Surface Nets | Quad | Yes | ~1x | Voxel Edge Weight | | Marching Cubes | Triangle | No/Partial | 1x | Linear on Edge | | Marching Tetrahedra | Triangle | Yes | 3x | Linear on Edge |

Meshing.MarchingCubes — Type.

MarchingCubes(iso=0.0, eps=1e-3, reduceverts=true, insidepositive=false)
MarchingCubes(;iso=0.0, eps=1e-3, reduceverts=true, insidepositive=false)
MarchingCubes(iso)
MarchingCubes(iso,eps)

Specifies the use of the Marching Cubes algorithm for isosurface extraction. This algorithm provides a good balance between performance and vertex count. In contrast to the other algorithms, vertices may be repeated, so face counts may be large.

iso (default: 0.0) specifies the iso level to use for surface extraction.

eps (default: 1e-3) is the tolerence around a voxel corner to ensure manifold mesh generation.

reduceverts (default: true) if true will merge vertices within a voxel to reduce mesh size by around 30% and with no performance penalty.

insidepositive (default: false) set true if the sign convention inside the surface is positive, common for NRRD and DICOM data

source

Meshing.MarchingTetrahedra — Type.

MarchingTetrahedra(iso=0.0, eps=1e-3, reduceverts=true, insidepositive=false)
MarchingTetrahedra(;iso=0.0, eps=1e-3, reduceverts=true, insidepositive=false)
MarchingTetrahedra(iso)
MarchingTetrahedra(iso,eps)

Specifies the use of the Marching Tetrahedra algorithm for isosurface extraction. This algorithm has a roughly 2x performance penalty compared to Marching Cubes, and generates more faces. However, each vertex is guaranteed to not be repeated, making this algorithm useful for topological analysis.

iso specifies the iso level to use for surface extraction.

eps is the tolerence around a voxel corner to ensure manifold mesh generation.

reduceverts reserved for future use.

insidepositive reserved for future use.

source

Meshing.NaiveSurfaceNets — Type.

NaiveSurfaceNets(iso=0.0, eps=1e-3, reduceverts=true, insidepositive=false)
NaiveSurfaceNets(;iso=0.0, eps=1e-3, reduceverts=true, insidepositive=false)
NaiveSurfaceNets(iso)
NaiveSurfaceNets(iso,eps)

Specifies the use of the Naive Surface Nets algorithm for isosurface extraction. This algorithm has a slight performance advantage in some cases over Marching Cubes. Each vertex is guaranteed to not be repeated, however the algorithm does not guarantee accuracy and generates quad faces.

iso specifies the iso level to use for surface extraction.

eps is the tolerence around a voxel corner to ensure manifold mesh generation.

reduceverts reserved for future use.

insidepositive reserved for future use.

source

Meshing.AbstractMeshingAlgorithm — Type.

AbstractMeshingAlgorithm

Abstract type to specify an algorithm for isosurface extraction. (unexported)

source

Isosurface

isosurface is the common and generic API for isosurface extraction with any type of abstract vector/vertex/face type.

Meshing.isosurface — Function.

`function isosurface(sdf::AbstractArray{T, 3}, method::AbstractMeshingAlgorithm,
                     [ VertType = SVector{3,Float64} ], [ FaceType} = SVector{3, Int} ] ;
                     origin = SVector(-1.0,-1.0,-1.0), widths = SVector(2.0,2.0,2.0))`

`function isosurface(f::Function, method::AbstractMeshingAlgorithm,
                     [ VertType = SVector{3,Float64} ], [FaceType = SVector{3, Int} ] ;
                     origin = SVector(-1.0,-1.0,-1.0), widths = SVector(2.0,2.0,2.0),
                     samples=(50,50,50))`

source

GeometryTypes

Meshing extends the mesh types in GeometryTypes for convience and use with visualization tools such as Makie and MeshCat.